+++ 2009-old/HwrSplitter/HwrLib/SymbolClass.h
@@ -5,17 +5,17 @@
 class SymbolClass
 {
 public:
-	Float mLength;
-	Float sLength;
-	Float wLength;
+	double mLength;
+	double sLength;
+	double wLength;
 	wchar_t originalChar;
 	CombinedFeatureDistribution phase[SUB_PHASE_COUNT];
-	SymbolClass(Float meanLength, Float varLength) : mLength(meanLength), wLength(100), sLength(varLength*100)	{ }
+	SymbolClass(double meanLength, double varLength) : mLength(meanLength), wLength(100), sLength(varLength*100)	{ }
 	SymbolClass() : mLength(0.0), wLength(0.0), sLength(0.0) {}
 	
-	Float meanLength() const {return mLength;}
-	Float varLength() const {return sLength/wLength;}
-	Float weightLength() const {return wLength;}
+	double meanLength() const {return mLength;}
+	double varLength() const {return sLength/wLength;}
+	double weightLength() const {return wLength;}
 
 	void ScaleWeightBy(double scaleFactor) {
 		wLength*=scaleFactor;
@@ -28,8 +28,8 @@ public:
 		return -0.5*sqr(length - mLength)/sLength*wLength; //we can ignore the DC offset - after all this is constant for any length.
 	}
 
-	void LearnLength(Float length, Float weight) {
-		Float newWeight = wLength + weight;
+	void LearnLength(double length, double weight) {
+		double newWeight = wLength + weight;
 		sLength = sLength + sqr(length - mLength)*wLength*weight/newWeight;
 		mLength = mLength + (length - mLength)*weight/newWeight;
 		wLength = newWeight;
