+++ 2009-old/HwrSplitter/HwrLib/image/drawing.h
@@ -18,9 +18,9 @@ void block_rgb(PamImage<RGBPixel>& im, int x1, int y1, int x2, int y2, RGBPixel 
 void horizontal_line_rgb(PamImage<RGBPixel>& im, int y, RGBPixel color, float alpha = 0.7);
 
 /// Draw a x-projection histogram
-void draw_x_projection(PamImage<RGBPixel>& im, int x0, int x1, const std::vector<Float>& histo, RGBPixel color, float alpha = 0.7);
+void draw_x_projection(PamImage<RGBPixel>& im, int x0, int x1, const std::vector<double>& histo, RGBPixel color, float alpha = 0.7);
 /// Draw a x-projection histogram
-void draw_y_projection(PamImage<RGBPixel>& im, int y0, int y1, const std::vector<Float>& histo, RGBPixel color, float alpha = 0.7);
+void draw_y_projection(PamImage<RGBPixel>& im, int y0, int y1, const std::vector<double>& histo, RGBPixel color, float alpha = 0.7);
 
 /// Draw a boundary
 void draw_boundary(PamImage<RGBPixel>& im, const Boundary& b);
